+++ app/main.py
@@ -1,7 +1,7 @@
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 
-from app.routers import users
+from app.routers import users, walks
 
 app = FastAPI(
     title="Open Forum",
@@ -27,3 +27,4 @@ app.add_middleware(
 )
 
 app.include_router(users.router, prefix="/api/v1", tags=["users"])
+app.include_router(walks.router, prefix="/api/v1", tags=["walks"])